HBX-1958: Rename class 'org.hibernate.tool.internal.reveng.BasicColumnProcessor' to 'org.hibernate.tool.internal.reveng.reader.BasicColumnProcessor'
